Новая методика карточки проекта

+++ back/src/main/java/ru/hrhappiness/repository/forUPC/ProjectType1Repository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import ru.hrhappiness.entitys.userProjectCard.ProjectType1;
 
 @Repository
-public interface ProjectType1Repository extends JpaRepository<ProjectType1,Integer> {
+public interface ProjectType1Repository extends JpaRepository<ProjectType1,Long> {
 }
